requirements.txt
certifi==2020.6.20
Django==2.2.14
djangorestframework==3.11.0
pkg-resources==0.0.0
pytz==2020.1
sqlparse==0.3.1
